Draft of enabling user to place ships
battleship.js now contains a load of functions that transfer the game
from the "Get Players" to "Player 1 - place ships" to player 1 actually
placing their ships.

Cleaned up the naming scheme of the HTML variables of the Player object.

Developed a plethora of functions in game-management.js to handle
adding/removing various pieces of the page (including but not limited
to the in-between message, orientation choice radio selection, and
placing ships).